apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
redis==5.0.4
httpx==0.27.0
bleach==6.1.0
slowapi==0.1.9
resend==0.7.2
stripe==9.9.0
python-dotenv==1.0.1
email-validator==2.1.1
Pillow==10.3.0
greenlet==3.0.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0
modal
